Fix YDA-4115
Size of revisions wasn't counted in statistics job because of
double slash in collection search path (UUREVISIONCOLLECTION already
starts with a slash, so doesn't need to be prefixed with a slash).

Also fixed potential issue where two group names with a common prefix
(e.g. research-jan and research-jansen) could result in a situation
where the contents of the group with the long name will be counted
towards the statistics of the group with the short name.