YDA-5416: improve robustness rev cleanup process
The solution involves a separate spooling system for batch jobs
that temporarily stores data for batch processing on disk. The
revision cleanup script now first triggers a rule to collect
all revision data, and stores this data in batches in the spooling
system. The cleanup script then runs rules that clean up revisions
in these batches.

The intent of this design is to minimize the impact of resource leaks in iRODS
(by processing revisions in batches) and to reduce the potential for
cascading failures (because batch processing now uses a separate
system for keeping track of batches, rather than a chain of rule
invocations).