YDA-4236 fix intake module prefix issues
When searching for a collection and its subcollections,
one needs to search for both "/zoneName/myCollection"
and "/zoneName/myCollection/%". Searching for '/zoneName/myCollection%"
is incorrect, since this will also match collections with the same
prefix (e.g. "/zoneName/myCollection2"), which can cause unexpected
results when processing collections with a common prefix.