YDA-5415: improve revision cleanup stability
This modifies the revision cleanup process, so that the scan stage
(collecting data about revisions) is separate from the collect stage
(collecting a list of all revisions). Scanning revision data is done
in batches now, rather than a one-time scan for the whole environment.

This reduces the amount of memory needed for the scan stage, which
contributes to making the revision cleanup job usable on large
environments.