Limit maximum memory usage in replication and revision job
Enable stopping when memory usage exceeds a configurable limit.

This prevents using too much memory when large amounts of data objects
need to be replicated or should get revisions and memory-leaks exist.

After the limit is reached, the process stops in a normal way.
Presumably they will be started again soon by the normal cron jobs or
whatever the operators use.

To facilitate testing the job can now also run in "dry_run" mode, where
it logs which data objects would be processed, but doesn't do so.

Note that the current memory leaking seems directly proportional to the
number of ICAT queries processed. So in dry_run you won't see everything.

The revisions job, which amongst others creates and copies AVU's is
especially prone to leaking memory very fast.

Ironically, especially in the replica job, checking whether a "stop"
data objects exists, is the major source of trouble.

In no_action mode the memory used is logged at each step.

This particular commit is more of a proof of concept, but aimed at being
ready to ad hoc deal with the memory issues. It specifically does not
address other issues that are present within these jobs.

Ad hoc:

    - limit set to 1.000.000.000 (1 GB), not configurable but easily
      temporarily adjusted with vi, as is the dry_run parameter
    - psutil code identical in both source files, not refactored to
      separate shared utility code yet
    - the "psutil" package is already present in current Yoda due to
      dependencies of the data access passwords code
    - [replica] and [revisions] tags in logging added to stay consistent
      with this version
    - manually tested, no unit or integration tests present
    - during testing several issues surfaced, some quite severe, but
      they all seemed already present and not related to the specific
      memory limiting in this patch