replication.py: refactor setting "org_replication_failed" job flag
It could fail when the user (normally rods) had no permission to access
the data object.

Changes were done in the current style. So rather verbose and with lots
of repetition.

There probably is a "rodsAdmin=" flag somewhere which could be used in
this case for iRODS version 4.2.12 and above, but for now this is at
least a practical improvement.