replication.py: calculate checksum as rodsAdmin and not with replNum=0
Although rare in current Yoda, it is not always the case that there is a
replNum 0 for a data object, nor that it has always a "good" status.

Due to the logic in the current replica.py, at this stage the user may
not yet have sufficient access to the data object. And anyway the
"rodsAdmin=" flag seems more appropriate.