replication.py: workaround for multiple org_replication_scheduled AVU's
Due to multiple PEP's firing and sometimes not correctly dealing with
them, it is possible to have multiple "org_replication_scheduled" AVU's
for the same data object, see pep_resource_modified_post.

On the first occurrence in the replica job (which processes each AVU one
by one, alphabetically sorted due to the query) the first AVU is processed
(either successfully or not) and then *all* AVU's are deleted.

So on the next occurrence in the loop within the replica job the extra
AVU's for the same data object are no longer physically present.

Just ignore them, but do report them for keeping track of the situation.

Once the, arguably, bug has been fixed that creates these multiple AVU's
the workaround can be removed in a later commit.

Although in a general iRODS setup, it certainly could be a valid scenario
that multiple replica's of the same data object could have been touched
while the replica job is running, and you would want to do something with
them. That is out of scope for now.

Instead of the "count" variable "len(data_objects_seen)" could now be
used, but I have left it as it is, only moving it after the "ignore"
part.