YDA-5783: fix logic for vault schemas
This fix implements additional logic for determining the active metadata
schema of a vault group.

In addition to the existing behaviour (using the metadata schema of
the matching deposit/research group), we now allow a schema_id AVU to be set on
the vault group that overrides the metadata schema. Furthermore, if no
vault schema_id AVU is set and no research/deposit group is found,
we revert to the default metadata schema of the environment.

This ensures that metadata conversion can work on vault groups that
no longer have a matching research/deposit group. If a standalone
vault group has a non-default metadata schema, its metadata schema
needs to be configured explicitly for it to be processed correctly
by the schema transformation rules.