YDA-5866 fix rev creation job stopping on error
The revision creation job used to stop with a failure when a data object scheduled for revision creation could no longer be found on the resource where the revision creation job was looking for it. This can happen e.g. after a resource tree change. The revision creation job now logs an error message and continues with processing the next data object scheduled for revision creation, so that one such error can't stop the entire revision creation process.

This also extracts the function to gather data object properties to util.data_object and covers it with integration tests, so that this function can be re-used in other parts of the ruleset.