YDA-6198: add all users internal option
Add a way to make the ruleset consider all users to
be internal, by setting the external domain filter to "*".
This is useful in situations where all users are to be authenticated
using OIDC.

This also makes the is-user-external script use the ruleset
function for determining whether a user is external, so that the
PAM stack and ruleset use exactly the same logic.