YDA-6524: skip unsupported packages in pub.update
When updating publication endpoints, skip data packages
with metadata schemas that are known to be unsupported. Trying
to generate publication data (e.g. landing pages) for such
data packages does not work and results in hard-to-read
log clutter, so it's better to just skip these packages with
an informational message.