YDA-6528: improve DM group error handling
Add error handling code for situations where we are unable
to determine the datamanager group of a research group, e.g.
because it has no category.

Research groups without a category are invalid, but they can
be created under various error conditions, so policies and
reporting jobs need to be able to handle such cases.