YDA-6460: add script to archive data collection
Shell script to use existing microservices to archive collections to a single file, and extract file to a collection, without need of upload and download step. If the archive extension is .zip, archives are also compressed.
Limitations: currently, target resource is set to be the default, specifying files for extraction is not supported, .zip extraction is not supported.

---------

Co-authored-by: claravox <claravox@users.noreply.github.com>